tests: These recently added tests require the testing repository
For an example of a failing build log, see:

https://travis-ci.org/sysrepo/sysrepo/jobs/385306547

fixes #1148
