+++ tests/rp_dt_edit_test.c
@@ -2590,201 +2590,6 @@ void set_and_get_item_id_ref(void **state){
     test_rp_session_cleanup(ctx, session);
 }
 
-static void set_item(const char *path, const char *value, sr_type_t type, rp_ctx_t *ctx,
-                     rp_session_t *session)
-{
-    sr_val_t *val = NULL;
-    int rc;
-    sr_new_val(path, &val);
-    if (type != SR_LIST_T) {
-        sr_val_set_str_data(val, type, value);
-    }
-    val->type = type;
-    SR_LOG_INF("Adding item %s", path);
-    assert_int_equal(val->type, type);
-
-    rc = rp_dt_set_item_wrapper(ctx, session, val->xpath, val, NULL, SR_EDIT_DEFAULT);
-    assert_int_equal(SR_ERR_OK, rc);
-}
-
-static int set_items_data_and_feature_import_setup(void **state)
-{
-    exec_shell_command("../src/sysrepoctl --install --yang=" TEST_SOURCE_DIR "/yang/data-feat-enable-A.yang", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-feat-enable-A.yang", true);
-    exec_shell_command("../src/sysrepoctl --install --yang=" TEST_SOURCE_DIR "/yang/data-feat-enable-B.yang", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-feat-enable-B.yang", true);
-    exec_shell_command("../src/sysrepoctl --install --yang=" TEST_SOURCE_DIR "/yang/data-feat-enable-C.yang", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-feat-enable-C.yang", true);
-    exec_shell_command("../src/sysrepoctl --install --yang=" TEST_SOURCE_DIR "/yang/data-feat-enable-D.yang", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-feat-enable-D.yang", true);
-
-    exec_shell_command("../src/sysrepoctl --feature-enable c-feature --module data-feat-enable-C", ".*", true, 0);
-
-    exec_shell_command("../src/sysrepoctl --install --yang=" TEST_SOURCE_DIR "/yang/data-imp-dep-A.yang", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-imp-dep-A.yang", true);
-    exec_shell_command("../src/sysrepoctl --install --yang=" TEST_SOURCE_DIR "/yang/data-imp-dep-B.yang", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-imp-dep-B.yang", true);
-    exec_shell_command("../src/sysrepoctl --install --yang=" TEST_SOURCE_DIR "/yang/data-imp-dep-C.yang", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-imp-dep-C.yang", true);
-
-    exec_shell_command("../src/sysrepoctl --install --yang=" TEST_SOURCE_DIR "/yang/data-submodule-main.yang", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-submodule-main.yang", true);
-
-    exec_shell_command("../src/sysrepoctl --install --yang=" TEST_SOURCE_DIR "/yang/feature-submodule-main.yang", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "feature-submodule-main.yang", true);
-
-    exec_shell_command("../src/sysrepoctl --feature-enable test-submodule-feature --module feature-submodule-main", ".*", true, 0);
-
-    setup(state);
-
-    return 0;
-}
-
-static int set_items_data_and_feature_import_teardown(void **state)
-{
-    exec_shell_command("../src/sysrepoctl --feature-disable c-feature --module data-feat-enable-C", ".*", true, 0);
-
-    exec_shell_command("../src/sysrepoctl --uninstall --module=data-feat-enable-A", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-feat-enable-A.yang", false);
-    exec_shell_command("../src/sysrepoctl --uninstall --module=data-feat-enable-B", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-feat-enable-B.yang", false);
-    exec_shell_command("../src/sysrepoctl --uninstall --module=data-feat-enable-C", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-feat-enable-C.yang", false);
-    exec_shell_command("../src/sysrepoctl --uninstall --module=data-feat-enable-D", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-feat-enable-D.yang", false);
-
-    exec_shell_command("../src/sysrepoctl --uninstall --module=data-imp-dep-A", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-imp-dep-A.yang", false);
-    exec_shell_command("../src/sysrepoctl --uninstall --module=data-imp-dep-B", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-imp-dep-B.yang", false);
-    exec_shell_command("../src/sysrepoctl --uninstall --module=data-imp-dep-C", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-imp-dep-C.yang", false);
-    exec_shell_command("../src/sysrepoctl --uninstall --module=data-submodule-main", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "data-submodule-main.yang", false);
-
-    exec_shell_command("../src/sysrepoctl --uninstall --module=feature-submodule-main", ".*", true, 0);
-    test_file_exists(TEST_SCHEMA_SEARCH_DIR "feature-submodule-main.yang", false);
-
-    teardown(state);
-
-    return 0;
-}
-
-void set_items_data_and_feature_import(void **state)
-{
-    int rc;
-    rp_ctx_t *ctx = *state;
-    rp_session_t *session = NULL;
-    sr_error_info_t *errors = NULL;
-    size_t e_cnt = 0;
-    dm_commit_context_t *c_ctx = NULL;
-
-    test_rp_session_create(ctx, SR_DS_STARTUP, &session);
-
-    set_item("/data-feat-enable-D:d-con/d-list[name='d-foo']", NULL, SR_LIST_T,
-             ctx, session);
-    set_item("/data-feat-enable-D:d-con/data-feat-enable-A:a-leaf", "a-bar", SR_STRING_T,
-             ctx, session);
-    set_item("/data-feat-enable-B:b-con/b-list[name='b-foo']", NULL, SR_LIST_T,
-             ctx, session);
-    set_item("/data-feat-enable-B:b-con/b-list[name='b-foo']/ref", "d-foo", SR_STRING_T,
-             ctx, session);
-
-
-    rc = rp_dt_commit(ctx, session, &c_ctx, false, &errors, &e_cnt);
-    assert_int_equal(rc, SR_ERR_OK);
-    assert_int_equal(e_cnt, 0);
-    assert_ptr_equal(errors, NULL);
-
-    test_rp_session_cleanup(ctx, session);
-}
-
-void set_items_data_and_import_implemented(void **state)
-{
-    int rc;
-    rp_ctx_t *ctx = *state;
-    rp_session_t *session = NULL;
-    sr_error_info_t *errors = NULL;
-    size_t e_cnt = 0;
-    dm_commit_context_t *c_ctx = NULL;
-
-    test_rp_session_create(ctx, SR_DS_STARTUP, &session);
-
-    set_item("/data-imp-dep-A:a-con/a-list[name='a-foo']", NULL, SR_LIST_T,
-             ctx, session);
-    set_item("/data-imp-dep-A:a-con/a-list[name='a-foo']/ref", "c-foo", SR_STRING_T,
-             ctx, session);
-    set_item("/data-imp-dep-C:c-con/c-list[name='c-foo']", NULL, SR_LIST_T,
-             ctx, session);
-    set_item("/data-imp-dep-C:c-con/data-imp-dep-B:b-leaf", "b-foo", SR_STRING_T,
-             ctx, session);
-
-
-    rc = rp_dt_commit(ctx, session, &c_ctx, false, &errors, &e_cnt);
-    assert_int_equal(rc, SR_ERR_OK);
-    assert_int_equal(e_cnt, 0);
-    assert_ptr_equal(errors, NULL);
-
-    test_rp_session_cleanup(ctx, session);
-}
-
-void set_and_get_items_from_top_level_in_submodule(void **state)
-{
-    int rc;
-    rp_ctx_t *ctx = *state;
-    rp_session_t *session = NULL;
-    sr_error_info_t *errors = NULL;
-    size_t e_cnt = 0;
-    dm_commit_context_t *c_ctx = NULL;
-    sr_val_t *val = NULL;
-
-    test_rp_session_create(ctx, SR_DS_STARTUP, &session);
-
-    set_item("/data-submodule-main:foo/bar", "test", SR_STRING_T, ctx, session);
-
-    rc = rp_dt_commit(ctx, session, &c_ctx, false, &errors, &e_cnt);
-    assert_int_equal(rc, SR_ERR_OK);
-    assert_int_equal(e_cnt, 0);
-    assert_ptr_equal(errors, NULL);
-
-    rc = rp_dt_get_value_wrapper(ctx, session, NULL, "/data-submodule-main:foo/bar", &val);
-    assert_int_equal(SR_ERR_OK, rc);
-    assert_non_null(val);
-    assert_int_equal(SR_STRING_T, val->type);
-    assert_string_equal("test", val->data.string_val);
-    sr_free_val(val);
-
-    test_rp_session_cleanup(ctx, session);
-}
-
-void set_and_get_items_from_feature_of_submodule(void **state)
-{
-    int rc;
-    rp_ctx_t *ctx = *state;
-    rp_session_t *session = NULL;
-    sr_error_info_t *errors = NULL;
-    size_t e_cnt = 0;
-    dm_commit_context_t *c_ctx = NULL;
-    sr_val_t *val = NULL;
-
-    test_rp_session_create(ctx, SR_DS_STARTUP, &session);
-    set_item("/feature-submodule-mod:mod-container/feature-submodule-main:augmented-data-val", "aug-data", SR_STRING_T, ctx, session);
-
-    rc = rp_dt_commit(ctx, session, &c_ctx, false, &errors, &e_cnt);
-    assert_int_equal(rc, SR_ERR_OK);
-    assert_int_equal(e_cnt, 0);
-    assert_ptr_equal(errors, NULL);
-
-    rc = rp_dt_get_value_wrapper(ctx, session, NULL, "/feature-submodule-mod:mod-container/feature-submodule-main:augmented-data-val", &val);
-    assert_int_equal(SR_ERR_OK, rc);
-    assert_non_null(val);
-    assert_int_equal(SR_STRING_T, val->type);
-    assert_string_equal("aug-data", val->data.string_val);
-    sr_free_val(val);
-
-    test_rp_session_cleanup(ctx, session);
-}
-
 int main(){
 
     sr_log_stderr(SR_LL_DBG);
@@ -2825,10 +2630,6 @@ int main(){
             cmocka_unit_test_setup(edit_union_type, createData),
             cmocka_unit_test_setup(validaton_of_multiple_models, createData),
             cmocka_unit_test(set_and_get_item_id_ref),
-            cmocka_unit_test_setup_teardown(set_items_data_and_feature_import, set_items_data_and_feature_import_setup, set_items_data_and_feature_import_teardown),
-            cmocka_unit_test_setup_teardown(set_items_data_and_import_implemented, set_items_data_and_feature_import_setup, set_items_data_and_feature_import_teardown),
-            cmocka_unit_test_setup_teardown(set_and_get_items_from_top_level_in_submodule, set_items_data_and_feature_import_setup, set_items_data_and_feature_import_teardown),
-            cmocka_unit_test_setup_teardown(set_and_get_items_from_feature_of_submodule, set_items_data_and_feature_import_setup, set_items_data_and_feature_import_teardown)
     };
 
     watchdog_start(300);
